chore: moved validator to a separate file
